Add Check for Blank Form Fields
A bit of is due with this one. Because the form starts with display:none
it is not initially rendered in the DOM. When the form submission button
was clicked and the HTML elements were `required`, the console would
throw an error: `An invalid form control with name='' is not focusable.`
To get around this, I tried to add the required when the add book
button was clicked, but this did not work because the change took effect
after the form was already rendered. So, I wrote a quick value check to
ensure the the author, title, and/or pages fields did not equal `' '`.
It may be a bit non-standard, but it works.