Fix CSS Color Variables
This one if humorous. I did a find and replace in VS Code to replace all
the manually declared colors with the CSS variables, but I also replaced
the value of the variables themselves doing this. So, I made
`--main: var(--main);`
That did not work. This commit fixes it.